requirements.txt
telethon==1.41.2
fastapi==0.99.1
uvicorn[standard]==0.22.0
sqlalchemy==2.0.43
databases==0.9.0
python-multipart==0.0.6
python-dotenv==1.0.0
Pillow==10.0.1
aiofiles==23.1.0
httpx==0.27.2

